requirements.txt
flask
mysql-connector-python
demoji
pythainlp